# server/swagger.yaml
openapi: 3.0.2
info:
  title: Api Rest Proyect
  description: API to manage users.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local server
paths:
  /clients:
    get:
      summary: Get all users
      description: Returns a list of all users in the database.
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new user
      description: Creates a new user in the database.
      requestBody:
        description: New user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: string
                example: "Inserted successfully"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        Nombre:
          type: string
          description: User's first name
        Apellido:
          type: string
          description: User's last name
        mail:
          type: string
          description: User's email address
    UserInput:
      type: object
      properties:
        Nombre:
          type: string
          description: User's first name
        Apellido:
          type: string
          description: User's last name
        mail:
          type: string
          description: User's email address
      required:
        - Nombre
        - Apellido
        - mail
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error description